Add files via upload
# Use Neural Networks to solve a classification task from
# https://archive.ics.uci.edu/ml/datasets.php. Use two methods/architectures/NN
# types.
We will use real-life datasets to demonstrate the basics of machine learning, which
include preprocessing data for machine learning models and building a classification
model using two architectures(scikit-learn and keras).
We will use a wine dataset
(https://archive.ics.uci.edu/ml/machine-learning-databases/wine/). These data are
the results of a chemical analysis of wines grown in the same region in Italy but 
derived from three different cultivars. The analysis determined the quantities of 13
constituents found in each of the three types of wines.